backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
openai>=1.57.0
anthropic>=0.40.0
google-generativeai>=0.8.0
python-multipart>=0.0.6
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
bcrypt>=4.1.2
